Last Update before presentation

+++ Assets/_Components/Cards/Scripts/CardData.cs
@@ -14,7 +14,7 @@ public class CardData
 
     public float range;
 
-    public int movementSpeed;
+    public float movementSpeed;
 
     //public int attackSpeed;
