+++ Assets/_Components/Cards/Scripts/CardSO.cs
@@ -14,7 +14,7 @@ public class CardSO : ScriptableObject
 
     public float range;
 
-    public int movementSpeed;
+    public float movementSpeed;
 
     public int attackSpeed;
